Added a Parser Interface for all the AddTaskParser, AddEventParser, ... classes The Switch Case is managed by ParserManager.
