+++ src/main/java/seedu/task/logic/LogicManager.java
@@ -3,7 +3,7 @@ package seedu.task.logic;
 import javafx.collections.ObservableList;
 import seedu.task.logic.commands.Command;
 import seedu.task.logic.commands.CommandResult;
-import seedu.task.logic.parser.Parser;
+import seedu.task.logic.parser.ParserManager;
 import seedu.task.model.Model;
 import seedu.task.model.item.ReadOnlyTask;
 import seedu.task.storage.Storage;
@@ -19,11 +19,11 @@ public class LogicManager extends ComponentManager implements Logic {
     private final Logger logger = LogsCenter.getLogger(LogicManager.class);
 
     private final Model model;
-    private final Parser parser;
+    private final ParserManager parser;
 
     public LogicManager(Model model, Storage storage) {
         this.model = model;
-        this.parser = new Parser();
+        this.parser = new ParserManager();
     }
 
     @Override
